Refactor disk_constraints into disks module
