+++ astoria/common/disks/__init__.py
@@ -1,0 +1,1 @@
+"""Code for recognising disk drives."""
